# Improve performance of filter_dict