--- a/arrow/src/compute/kernels/filter.rs
+++ b/arrow/src/compute/kernels/filter.rs
@@ -786,7 +786,13 @@ where
         )
     };
 
-    DictionaryArray::<T>::from(data)
+    unsafe {
+        DictionaryArray::<T>::try_new_unchecked(
+            filtered_keys,
+            array.values().clone(),
+            data,
+        )
+    }
 }
 
 #[cfg(test)]